requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
hf-xet==1.1.5
huggingface-hub==0.33.4
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.2
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai-whisper==20250625
optuna==4.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
primePy==1.3
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.41
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.7.4
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
